requirement.txt
python-telegram-bot
python-dotenv
pymongo
langchain
openai